cilium service mesh is used to make sure the proxy service is used by ActiveGate
